codecs: silence NDK clang strict-warning failures
Two warnings that Clang 21 (NDK r30) elevates to errors under our
-Werror -Wsign-conversion -Wunused-function set, but desktop
GCC 13 / Clang 17 didn't fire on:

  atari_seq/decoders.cc:122
    `(d2 << inv_shift) & 0xFFu`. d2 is uint8_t; integer-promotion
    upgrades it to signed int before the shift, then `& 0xFFu`
    forces a signed-to-unsigned conversion. Promote d2 to unsigned
    first — same shape as the shifted_d2 line above.

  yafa/decoders.cc:10
    Anonymous-namespace `fourcc` helper was never called in this
    TU. NDK clang flags it via -Wunused-function. Drop it; if a
    user of the kXPKF literal pattern needs it, it's two lines to
    bring back.

+++ codecs/src/atari_seq/src/decoders.cc
@@ -119,7 +119,11 @@ namespace atari_seq {
             const auto shifted_d2 =
                 static_cast <std::uint8_t>(((d1 << inv_shift) & inv_mask)
                                            | static_cast <unsigned>(d2 >> shift));
-            const auto shifted_d3 = static_cast <std::uint8_t>((d2 << inv_shift) & 0xFFu);
+            // Promote d2 to unsigned before the shift so the `& 0xFFu`
+            // mask doesn't trigger -Wsign-conversion (uint8_t integer-
+            // promotes to signed int by default; matches the pattern on
+            // the shifted_d2 line above).
+            const auto shifted_d3 = static_cast <std::uint8_t>((static_cast <unsigned>(d2) << inv_shift) & 0xFFu);
 
             if (xor_op) {
                 w.fb[w.si] = static_cast <std::uint8_t>(
